# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Server pets",
        "title": "Swagger Example API",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/playListVideo/createPlayListVideo": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PlayListVideo"
                ],
                "summary": "创建PlayListVideo",
                "parameters": [
                    {
                        "description": "创建PlayListVideo",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/video.PlayListVideo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"获取成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/playListVideo/deletePlayListVideo": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PlayListVideo"
                ],
                "summary": "删除PlayListVideo",
                "parameters": [
                    {
                        "description": "删除PlayListVideo",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/video.PlayListVideo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"删除成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/playListVideo/deletePlayListVideoByIds": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PlayListVideo"
                ],
                "summary": "批量删除PlayListVideo",
                "parameters": [
                    {
                        "description": "批量删除PlayListVideo",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.IdsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"批量删除成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/playListVideo/findPlayListVideo": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PlayListVideo"
                ],
                "summary": "用id查询PlayListVideo",
                "parameters": [
                    {
                        "type": "string",
                        "name": "auditReason",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "auditedAt",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "创建时间",
                        "name": "createdAt",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "createdBy",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "deletedBy",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "主键ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "playListId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "prodAuditReason",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "prodAuditedAt",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "prodId",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "prodStatus",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "prodUserId",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "更新时间",
                        "name": "updatedAt",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "updatedBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "url",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "userId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "version",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"查询成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/playListVideo/getPlayListVideoList": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PlayListVideo"
                ],
                "summary": "分页获取PlayListVideo列表",
                "parameters": [
                    {
                        "type": "string",
                        "name": "auditReason",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "auditedAt",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "创建时间",
                        "name": "createdAt",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "createdBy",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "deletedBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "endCreatedAt",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "主键ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "关键字",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页大小",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "playListId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "prodAuditReason",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "prodAuditedAt",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "prodId",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "prodStatus",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "prodUserId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "startCreatedAt",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "更新时间",
                        "name": "updatedAt",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "updatedBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "url",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "userId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "version",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"获取成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/playListVideo/updatePlayListVideo": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PlayListVideo"
                ],
                "summary": "更新PlayListVideo",
                "parameters": [
                    {
                        "description": "更新PlayListVideo",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/video.PlayListVideo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"更新成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.IdsReq": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "video.PlayListVideo": {
            "type": "object",
            "properties": {
                "auditReason": {
                    "type": "string"
                },
                "auditedAt": {
                    "type": "string"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "createdBy": {
                    "type": "integer"
                },
                "deletedBy": {
                    "type": "integer"
                },
                "id": {
                    "description": "主键ID",
                    "type": "integer"
                },
                "playListId": {
                    "type": "integer"
                },
                "prodAuditReason": {
                    "type": "string"
                },
                "prodAuditedAt": {
                    "type": "string"
                },
                "prodId": {
                    "type": "integer"
                },
                "prodStatus": {
                    "type": "boolean"
                },
                "prodUserId": {
                    "type": "integer"
                },
                "status": {
                    "type": "boolean"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                },
                "updatedBy": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                },
                "version": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-token",
            "in": "header"
        }
    }
}